test: disable auto-generated snapshot stories
They currently don't add much value, and so using Chromatic snapshot credits on them seems like a
waste. We should consider using the code for generating the variants more explicitly in stories if
we want to test multiple variants in the future. That way, we can have one story (and thus one
snapshot) for all the variants together.