chore: fix GitHub release changelog by using a workspace changelog
After changing `release.projectsRelationship` to `"fixed"`, only one version tag and GitHub release
was generated. But since we didn't generate a workspace changelog, the first project changelog
(`@udir-design/theme`) would be used. Since that changelog only contains changes that affect the
theme, it would not corrctly reflect the changes in the overall release.

Using a workspace changelog fixes this issue.