build: ensure test projects properly typecheck
Because our test projects use React 18, and resolve our main libraries through
the `workspace:` protocol, we ran into the issue with different versions of
peer dependencies which [pnpm describes in their docs](https://pnpm.io/package_json#dependenciesmetainjected)

This is solved by specifying `@udir-design/react` as "injected" in the test apps,
which ensures that a copy of the library which depends on `react@18` is used instead
of just symlinking into the lib directory where `react@19` is installed.

An additional problem was that `@navikt/aksel-icons` implicitly referenced
`@types/react@19` due to `react` not being specified as a (peer) dependency at all.
This is solved with `packageExtensions` in `pnpm-workspace.yaml`, so that the test apps
correctly reference a version of the library which depends on `react@18`.