build: ensure Chromatic can detect css-only changes to components
Before this, changing only a component's `.css` file would not trigger
a new snapshot for that component's stories in Chromatic. This was because
the component css was never imported by the `@udir-design/react` project itself.

We could have added `"../css/dist/**"` to `"externals"` in `chromatic.config.json`,
but this would cause **all** components to be re-snapshotted even if we only changed
a single component's css.

The new solution ensures that component css is imported from the component file, since storybook
now doesn't actually load the bundled `@udir-design/css` styling. Instead, a Vite plugin
transforms the `.css` source files so that, when Vite is run by Storybook, the imported styling
is included and wrapped in the correct css layer. However, when Vite is **not** run by Storybook,
the component `.css` files are stripped from the bundle. They are still included by
`@udir-design/css/src/index.css`, as before.

As a bonus, we now don't have to run the css build in watch mode concurrently with Storybook,
so the Storybook CLI console gets some more terminal real estate when run through Nx.